Copy main stylsheets from govuk_template
Enough to make the page render more or less the same.

Pretty much straight copied, but with some imports brought inline,
and the colours from `styleguide` moved to the relevant module.

Added a root CSS file for all template styles, mirrors existing setup
but might want this to be flexible in the future, part of app sass?

Added a `to-sort` folder to the ITCSS structure, as I was spending too
long thinking where things should go. Modules like accessibility need
splitting into many different layers of ITCSS, so skip that for now.

Doesn't include:
- No IE specific CSS
- No IE fonts
- No print