Add template for a alpha-beta component
alpha-beta is a good example of a simple component, comprised of
a template and SCSS, with explicit parameters for the template.

This structure, a single template with two roo CSS selectors isn't
how a proper component would be structured, instead we'd likely have
a `phase-banner` component, with a single template and modifiers in
the SASS for each phase.